Update XComStrategyTuning.ini
Includes all squad upgrade costs and individual weapon costs for Laser and Beam weapons for Legend difficulty. Also includes SMG squad upgrade costs for Magnetic and Beam weapons in Legend. 

Does not include individual weapon weapon costs for Magnetic and Beam weapons, as that is not included in this weapon pack.