Added XComGameData config for easier testing
When you launch a debug campaign from the developer shell, DEBUG_StartingItems array will automatically add all items defined in it to the players inventory. I didnt change the default amount of 10 for all items, but removed the Chosen guns for now since they tend to make it very hard to select weapons for testing.